requirements.txt
fastapi~=0.115.12
uvicorn
jinja2~=3.1.6
python-multipart
openai
python-dotenv~=1.1.0
pydantic~=2.11.6
aiofiles~=24.1.0
langchain~=0.3.25
pytest~=8.4.0
pytest-asyncio
langchain-community~=0.3.25
langchain-openai~=0.3.23
langchain-core~=0.3.65
pillow~=11.2.1
duckduckgo-search
tiktoken~=0.9.0
sentence-transformers~=4.1.0
pytest-cov